refactor: backend folder structure
