+++ backend/package.json
@@ -3,7 +3,9 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon index.js",
+    "dev": "nodemon src/local.js",
+    "start": "node src/local.js",
+    "vercel": "node src/server.js",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
   },
