+++ backend/src/redis/redisClient.js
@@ -1,5 +1,5 @@
 import { Redis } from "@upstash/redis";
-import { ENV } from "./config.js";
+import { ENV } from "../config/env.js";
 
 export const redis = new Redis({
   url: ENV.REDIS_URL,
