+++ backend/src/server.js
@@ -1,0 +1,2 @@
+import { app } from "./app.js";
+export default app;
