+++ backend/src/services/radioService.instance.js
@@ -1,6 +1,6 @@
 import fetch from "node-fetch";
-import { db } from "./db.js";
-import { config } from "./config.js";
+import { db } from "../db/db.js";
+import { config } from "../config/config.js";
 import { createRadioService } from "./radioService.js";
 
 export const radioService = createRadioService({
