+++ backend/src/utils/errorHandler.js
@@ -1,0 +1,6 @@
+export default function errorHandler(err, req, res, next) {
+  console.error("Error:", err);
+  res.status(err.status || 500).json({
+    message: err.message || "Server error",
+  });
+}
